% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSomatic.R
\name{MergeSomatic}
\alias{MergeSomatic}
\title{MergeSomatic}
\usage{
MergeSomatic(
  vcf1,
  vcf2,
  sources,
  GENO,
  id_t = "TUMOR",
  id_n = "NORMAL",
  pass_only = FALSE
)
}
\arguments{
\item{vcf1}{VCF object or vcf file}

\item{vcf2}{VCF object or vcf file}

\item{sources}{The caller of the VCF files.}

\item{GENO}{The GENO item sources of the mereged VCF. For example
c(GT = 1, AD = 2). The vector names are the GENO names and the
values are the corresponding index.}

\item{id_t}{The tumor sample ID.}

\item{id_n}{The normal/control sample ID.}

\item{pass_only}{If TRUE, only PASS variants will be merged.}
}
\value{
A merged VCF object
}
\description{
To merge two VCFs from different somatic callers.
}
