% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka2.R
\name{strelka_indel}
\alias{strelka_indel}
\title{Add GT, AD to strelka2 indel calls}
\usage{
strelka_indel(vcf)
}
\arguments{
\item{vcf}{The Indel VCF object from VariantAnnotation by strelka2.}
}
\value{
VCF object.
}
\description{
Add GT, AD to strelka2 indel calls
}
