% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka2.R
\name{strelka_snv}
\alias{strelka_snv}
\title{Add GT, AD to strelka2 snv calls}
\usage{
strelka_snv(vcf)
}
\arguments{
\item{vcf}{The SNV VCF object from VariantAnnotation by strelka2.}
}
\value{
VCF object.
}
\description{
Add GT, AD to strelka2 snv calls
}
