% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2sj.R
\name{vcf2sj}
\alias{vcf2sj}
\title{annotated VCF to SJ format}
\usage{
vcf2sj(avcf, tidx = 1, nidx = 2)
}
\arguments{
\item{avcf}{A vcf from VEP.}

\item{tidx}{The tumor sample index.}

\item{nidx}{The normal sample index.}
}
\description{
annotated VCF to SJ format
}
